AI Update
changed AI walk speed sp that they are not faster than the player.
And lowered damage done to player.

Fix AI HitBoxes so now registers the players capsule component and not any capsule.

Player's sword will not do damage on contact if sword is not drawn.
Later must be changed to only when th player attacks.